# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gestor de Estoque API",
    "description": "API REST para gestão de empresas, usuários, produtos, combos, promoções e movimentações de estoque.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Ambiente local" }
  ],
...
  "paths": {
...
    "/api/usuarios/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/id" }],
      "get": {
        "tags": ["Usuários"],
        "summary": "Buscar usuário por ID (somente da própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Usuário encontrado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Usuario" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["Usuários"],
        "summary": "Atualizar usuário (somente da própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UsuarioUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Usuario" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Usuários"],
        "summary": "Remover usuário (inativa, somente da própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/Message" },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/produtos": {
      "get": {
        "tags": ["Produtos"],
        "summary": "Listar produtos da própria empresa",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de produtos",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Produto" } } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Produtos"],
        "summary": "Criar produto",
        "description": "O empresaId é sempre o da empresa do usuário autenticado; qualquer valor enviado no corpo é ignorado.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProdutoCreate" } } }
        },
        "responses": {
          "201": {
            "description": "Produto criado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Produto" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/produtos/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/id" }],
      "get": {
        "tags": ["Produtos"],
        "summary": "Buscar produto por ID (somente da própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Produto encontrado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Produto" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["Produtos"],
        "summary": "Atualizar produto (somente da própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProdutoUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Produto atualizado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Produto" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Produtos"],
        "summary": "Remover produto (soft delete, somente da própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/Message" },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/promocoes": {
      "get": {
        "tags": ["Promoções"],
        "summary": "Listar promoções da própria empresa",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de promoções",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Promocao" } } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Promoções"],
        "summary": "Criar promoção",
        "description": "O empresaId é sempre o da empresa do usuário autenticado; qualquer valor enviado no corpo é ignorado.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PromocaoCreate" } } }
        },
        "responses": {
          "201": {
            "description": "Promoção criada",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Promocao" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/promocoes/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/id" }],
      "get": {
        "tags": ["Promoções"],
        "summary": "Buscar promoção por ID (somente da própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Promoção encontrada",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Promocao" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["Promoções"],
        "summary": "Atualizar promoção (somente da própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PromocaoUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Promoção atualizada",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Promocao" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Promoções"],
        "summary": "Remover promoção (somente da própria empresa)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/Message" },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/movimentacoes": {
      "get": {
        "tags": ["Movimentações"],
        "summary": "Listar movimentações da própria empresa",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de movimentações",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/MovimentacaoEstoque" } } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Movimentações"],
        "summary": "Registrar movimentação (ajusta o estoque do produto)",
        "description": "O usuarioId é sempre o do usuário autenticado; o produto precisa pertencer à empresa do usuário autenticado.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MovimentacaoCreate" } } }
        },
        "responses": {
          "201": {
            "description": "Movimentação registrada",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MovimentacaoEstoque" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token obtido em POST /api/auth/login. Envie como 'Authorization: Bearer <token>'."
      }
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "responses": {
      "Message": {
        "description": "Mensagem de confirmação",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } }
      },
      "Error": {
        "description": "Erro",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
      "Message": {
        "type": "object",
        "properties": { "message": { "type": "string" } }
      },
      "Error": {
        "type": "object",
        "properties": { "message": { "type": "string" } }
      },
...
      "EmpresaResumo": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "slug": { "type": "string" }
        }
      },
...
      "Usuario": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["ADMIN"] },
          "ativo": { "type": "boolean" },
          "empresaId": { "type": "string", "format": "uuid" },
          "empresa": { "$ref": "#/components/schemas/EmpresaResumo" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "UsuarioUpdate": {
        "type": "object",
        "properties": {
          "nome": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "senha": { "type": "string", "format": "password" },
          "ativo": { "type": "boolean" }
        }
      },
      "ProdutoImagem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "produtoId": { "type": "string", "format": "uuid" },
          "url": { "type": "string" },
          "alt": { "type": "string", "nullable": true },
          "ordem": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Produto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "empresaId": { "type": "string", "format": "uuid" },
          "codigo": { "type": "string" },
          "nome": { "type": "string" },
          "descricao": { "type": "string", "nullable": true },
          "precoVarejo": { "type": "number", "format": "double" },
          "precoAtacado": { "type": "number", "format": "double" },
          "estoque": { "type": "integer" },
          "fotoCapa": { "type": "string", "nullable": true },
          "ordemCatalogo": { "type": "integer" },
          "destaque": { "type": "boolean" },
          "visivelCatalogo": { "type": "boolean" },
          "ativo": { "type": "boolean" },
          "imagens": { "type": "array", "items": { "$ref": "#/components/schemas/ProdutoImagem" } },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "deletedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "ProdutoCreate": {
        "type": "object",
        "required": ["empresaId", "codigo", "nome", "precoVarejo", "precoAtacado"],
        "properties": {
          "empresaId": { "type": "string", "format": "uuid" },
          "codigo": { "type": "string" },
          "nome": { "type": "string" },
          "descricao": { "type": "string" },
          "precoVarejo": { "type": "number", "format": "double" },
          "precoAtacado": { "type": "number", "format": "double" },
          "estoque": { "type": "integer", "default": 0 },
          "fotoCapa": { "type": "string" },
          "ordemCatalogo": { "type": "integer", "default": 0 },
          "destaque": { "type": "boolean", "default": false },
          "ativo": { "type": "boolean", "default": true },
          "visivelCatalogo": { "type": "boolean", "default": true }
        }
      },
      "ProdutoUpdate": {
        "type": "object",
        "properties": {
          "codigo": { "type": "string" },
          "nome": { "type": "string" },
          "descricao": { "type": "string" },
          "precoVarejo": { "type": "number", "format": "double" },
          "precoAtacado": { "type": "number", "format": "double" },
          "estoque": { "type": "integer" },
          "fotoCapa": { "type": "string" },
          "ordemCatalogo": { "type": "integer" },
          "destaque": { "type": "boolean" },
          "ativo": { "type": "boolean" },
          "visivelCatalogo": { "type": "boolean" }
        }
      },
      "ComboItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "comboId": { "type": "string", "format": "uuid" },
          "produtoId": { "type": "string", "format": "uuid" },
          "quantidade": { "type": "integer" },
          "produto": { "$ref": "#/components/schemas/Produto" }
        }
      },
      "Combo": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "empresaId": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "descricao": { "type": "string", "nullable": true },
          "preco": { "type": "number", "format": "double" },
          "fotoCapa": { "type": "string", "nullable": true },
          "ordemCatalogo": { "type": "integer" },
          "destaque": { "type": "boolean" },
          "visivel": { "type": "boolean" },
          "ativo": { "type": "boolean" },
          "itens": { "type": "array", "items": { "$ref": "#/components/schemas/ComboItem" } },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "deletedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
      "PromocaoItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "promocaoId": { "type": "string", "format": "uuid" },
          "produtoId": { "type": "string", "format": "uuid", "nullable": true },
          "comboId": { "type": "string", "format": "uuid", "nullable": true },
          "preco": { "type": "number", "format": "double" },
          "produto": { "$ref": "#/components/schemas/Produto" },
          "combo": { "$ref": "#/components/schemas/Combo" }
        }
      },
      "Promocao": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "empresaId": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "dataInicio": { "type": "string", "format": "date-time" },
          "dataFim": { "type": "string", "format": "date-time" },
          "itens": { "type": "array", "items": { "$ref": "#/components/schemas/PromocaoItem" } },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "PromocaoItemInput": {
        "type": "object",
        "required": ["preco"],
        "description": "Informe produtoId OU comboId.",
        "properties": {
          "produtoId": { "type": "string", "format": "uuid" },
          "comboId": { "type": "string", "format": "uuid" },
          "preco": { "type": "number", "format": "double" }
        }
      },
      "PromocaoCreate": {
        "type": "object",
        "required": ["empresaId", "nome", "dataInicio", "dataFim", "itens"],
        "properties": {
          "empresaId": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "dataInicio": { "type": "string", "format": "date-time" },
          "dataFim": { "type": "string", "format": "date-time" },
          "itens": { "type": "array", "items": { "$ref": "#/components/schemas/PromocaoItemInput" } }
        }
      },
      "PromocaoUpdate": {
        "type": "object",
        "properties": {
          "nome": { "type": "string" },
          "dataInicio": { "type": "string", "format": "date-time" },
          "dataFim": { "type": "string", "format": "date-time" }
        }
      },
      "UsuarioResumo": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "nome": { "type": "string" },
          "email": { "type": "string", "format": "email" }
        }
      },
      "MovimentacaoEstoque": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "produtoId": { "type": "string", "format": "uuid" },
          "usuarioId": { "type": "string", "format": "uuid" },
          "tipo": { "type": "string", "enum": ["ENTRADA", "SAIDA", "AJUSTE"] },
          "quantidade": { "type": "integer" },
          "motivo": { "type": "string", "nullable": true },
          "produto": { "$ref": "#/components/schemas/Produto" },
          "usuario": { "$ref": "#/components/schemas/UsuarioResumo" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "MovimentacaoCreate": {
        "type": "object",
        "required": ["produtoId", "tipo", "quantidade"],
        "description": "usuarioId não é aceito no corpo: é sempre o usuário autenticado.",
        "properties": {
          "produtoId": { "type": "string", "format": "uuid" },
          "tipo": {
            "type": "string",
            "enum": ["ENTRADA", "SAIDA", "AJUSTE"],
            "description": "ENTRADA soma, SAIDA subtrai (erro se ficar negativo), AJUSTE define o estoque exatamente."
          },
          "quantidade": { "type": "integer" },
          "motivo": { "type": "string" }
        }
      }
    }
  }
}